Port the IO subsystem to the SDL2 port (issue #516)
- BreaknesSDL: IOProcessor (IOProcessing.h/.cpp) that turns SDL2 Input
  events (keyboard keys, game controller buttons/sticks/triggers) into
  events and dispatches them to the devices attached to the motherboard
  via bindings, like the managed IOProcessor (IOProcessing.cs).
- BreaknesSDL: IO settings (IOConfig.h/.cpp) stored in IOConfigSDL.json -
  separately from the managed application (IOConfig.json) so they never
  overlap; the format repeats the managed config.
- BreaknesSDL: interactive IO configurator (IOConfigurator.h/.cpp) based
  on the logic of the managed FormIOConfig dialog: device pool (add/
  remove), actuator bindings (wait for an input event), attach/detach
  to the motherboard ports. Launched with 'breaknes --ioconfig [path]'.
- Device identifiers are taken from the IO::DeviceID enum (IO/IO.h)
  instead of hardcoded values.
- main.cpp: --ioconfig command line parameter; attach configured devices
  at startup and forward SDL events to the IO processor in the event loop.
- CMakeLists.txt/vcxproj: add the new sources; add CartPcb/CartPcbPort.cpp
  missing from the CMake build since the issue #509 refactor.
- JsonLib: include <cstring> and use %llu for Int serialization so the
  Linux build compiles and produces sane JSON (the SDL port is the first
  consumer of the serializer).

+++ Common/JsonLib/Json.cpp
@@ -755,7 +755,7 @@ void Json::Value::Serialize(SerializeContext* ctx, int depth, bool sizeOnly)
 			Json::EmitText(ctx, value.AsBool ? "true" : "false", sizeOnly);
 			break;
 		case ValueType::Int:
-			swprintf(temp, sizeof(temp) / sizeof(temp[0]) - 1, L"%I64u", value.AsInt);
+			swprintf(temp, sizeof(temp) / sizeof(temp[0]) - 1, L"%llu", (unsigned long long)value.AsInt);
 			EmitWcharString(ctx, temp, sizeOnly);
 			break;
 		case ValueType::Float:
